requirements.txt
# FastAPI y dependencias
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6

# Firebase
firebase-admin==6.3.0

# AWS S3 (para almacenamiento de fotos DAGMA)
boto3==1.34.0

# Prometheus para métricas
prometheus-client==0.19.0

# Rate limiting
slowapi==0.1.9

# Utilidades
python-dotenv==1.0.0

# Testing (desarrollo)
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2
